third_party/blink/web_tests/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-base-appearance-computed-style-expected.txt

This is a testharness.js-based test.
[FAIL] UA styles of base appearance option::checkmark.
  assert_equals: content expected "\\"✓\\" / \\"\\"" but got "\\"✓\\""
Harness: the test ran to completion.
